requirements.txt
requests~=2.32
beautifulsoup4~=4.12
pandas~=2.2
sqlalchemy~=2.0
google-auth ~=2.36
google-api-python-client ~=2.152
pytest~=7.4
pytest-cov~=4.1
coverage~=7.3